requirements.txt
asgiref==3.8.1
certifi==2025.1.31
cffi==1.17.1
charset-normalizer==3.4.1
cryptography==44.0.2
Django==5.1.7
django-allauth==65.7.0
django-filter==25.1
django-oauth-toolkit==3.0.1
djangorestframework==3.16.0
djangorestframework_simplejwt==5.5.0
idna==3.10
jwcrypto==1.5.6
oauthlib==3.2.2
psycopg2-binary==2.9.10
pycparser==2.22
PyJWT==2.9.0
requests==2.32.3
shortuuid==1.0.13
sqlparse==0.5.3
typing_extensions==4.13.0
urllib3==2.3.0
